+++ test/index.js
@@ -1,6 +1,6 @@
-require('./board');
-require('./conveyor-area');
-require('./get-lane-starts.js');
-require('./lane');
-require('./loading-area');
-require('./player-events');
+require('./board-test');
+require('./conveyor-area-test');
+require('./get-lane-starts-test');
+require('./lane-test');
+require('./loading-area-test');
+require('./player-events-test');
