+++ test/loading-area-test.js
