+++ test/player-events-test.js
